LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text-styles/2d.text.draw.baseline.top-expected.txt

2d.text.draw.baseline.top
textBaseline top is the top of the em square (not the bounding box)
Actual output:
Expected output:


FAIL textBaseline top is the top of the em square (not the bounding box) assert_approx_equals: Red channel of the pixel at (5, 5) expected 0 +/- 2 but got 255
